knative/postprocess/requirements.txt
Flask==2.0.3
redis==4.3.4
psycopg2-binary==2.9.3
requests==2.26.0
gunicorn==20.1.0